Folder refactoring and add game description and instructions in README.md
