+++ README.md
@@ -1,3 +1,15 @@
-FULL ARKANOID GAME CODE
+"ARKANOID"
 
-Hello AT guys, feel free to learn about the code, test it and run it.
+Description:
+You control a small platform (Paddle), which prevents a ball from leaving the playing area, making it bounce. At the top are bricks, which disappear when touched by the ball.
+
+This game has 3 levels, by overcoming each level the next one increases the difficulty by adding one more row.
+
+You have 3 lives at the beginning, and after you lose them all, you will have to start over from the beginning.
+
+Instructions:
+-To start the game you can press the "up button or w".
+-To move the paddle you can use the direction buttons: "left and right" or "a and d"
+-To restart the game you must click on "Play again!"
+
+Have fun! :)
