+++ index.html
@@ -22,6 +22,6 @@
     </div>
 
 	<canvas id="breakout" width="400" height="500"></canvas>
-	<script type="module" src="/src/script.js" ></script>	
+	<script type="module" src="/src/main.js" ></script>	
 </body>
 </html>
