+++ src/actions/drawPaddle.js
