+++ src/actions/drawStage.js
